requirements.txt
amqp==2.5.1
backports.csv==1.0.7
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.4
django-import-export==1.2.0
django-rest-passwordreset==1.0.0
django-stubs==1.1.0
djangorestframework==3.10.2
djangorestframework-stubs==1.0.0
et-xmlfile==1.0.1
idna==2.8
importlib-metadata==0.23
jdcal==1.4.1
kombu==4.6.5
more-itertools==7.2.0
mypy==0.720
mypy-extensions==0.4.1
odfpy==1.4.0
openpyxl==2.6.3
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
sentry-sdk==0.11.2
sqlparse==0.3.0
tablib==0.13.0
typed-ast==1.4.0
typing-extensions==3.7.4
ujson==1.35
urllib3==1.25.3
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
